+++ src/main/java/org/bytedeco/javacpp/tools/Context.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2015 Samuel Audet
+ * Copyright (C) 2014-2019 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -52,6 +52,7 @@ class Context {
     String cppName = null;
     String javaName = null;
     boolean inaccessible = false;
+    boolean objectify = false;
     boolean virtualize = false;
     Declarator variable = null;
     InfoMap infoMap = null;