+++ src/main/java/org/bytedeco/javacpp/tools/Info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2018 Samuel Audet
+ * Copyright (C) 2014-2019 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -51,6 +51,7 @@ public class Info {
         cast = i.cast;
         define = i.define;
         flatten = i.flatten;
+        objectify = i.objectify;
         translate = i.translate;
         skip = i.skip;
         purify = i.purify;
@@ -91,6 +92,8 @@ public class Info {
     /** Outputs declarations for this class into their subclasses as well.
      * Also adds methods for explicit casting, as done for multiple inheritance by default. */
     boolean flatten = false;
+    /** Map global functions to instance methods, without {@code static} modifier, to implement an interface, etc. */
+    boolean objectify = false;
     /** Attempts to translate naively the statements of variable-like macros to Java. */
     boolean translate = false;
     /** Skips entirely all the code associated with the C++ identifiers, expressions, or header filenames.
@@ -124,6 +127,8 @@ public class Info {
     public Info enumerate(boolean enumerate) { this.enumerate = enumerate; return this; }
     public Info flatten() { this.flatten = true; return this; }
     public Info flatten(boolean flatten) { this.flatten = flatten; return this; }
+    public Info objectify() { this.objectify = true; return this; }
+    public Info objectify(boolean objectify) { this.objectify = objectify; return this; }
     public Info translate() { this.translate = true; return this; }
     public Info translate(boolean translate) { this.translate = translate; return this; }
     public Info skip() { this.skip = true; return this; }